Move and update raw block importing function.
